新增api

+++ src/api/controller/tool.js
@@ -13,6 +13,14 @@ export default class extends think.controller.rest {
         super.init(http);
     }
 
+    /**
+     * before magic method
+     * @return {Promise} []
+     */
+    __before() {
+
+    }
+
 
     /**
      * 查询数据
@@ -21,11 +29,14 @@ export default class extends think.controller.rest {
         let data;
         if (this.id) {
             let pk = await this.modelInstance.getPk();
-            data = await this.modelInstance.where({ [pk]: this.id }).find();
+            data = await this.modelInstance.where({[pk]: this.id}).find();
             return this.success(data);
         }
-        let values = this.get();
-        data = await this.modelInstance.page(values.page, values.pageSize || 10).countSelect();
+        if(this.get('user_id')){
+            data = await this.modelInstance.where({user_id: this.get('user_id')}).select();
+            return this.success(data);
+        }
+        data = await this.modelInstance.select();
         return this.success(data);
     }
 
@@ -48,22 +59,28 @@ export default class extends think.controller.rest {
      * 删除数据
      */
     async deleteAction() {
+        if (!this.id) {
+            return this.fail('params error');
+        }
         let pk = await this.modelInstance.getPk();
-        let rows = await this.modelInstance.where({ [pk]: this.id }).delete();
-        return this.success({ affectedRows: rows });
+        let rows = await this.modelInstance.where({[pk]: this.id}).delete();
+        return this.success({affectedRows: rows});
     }
 
     /**
      * 更新数据
      */
     async putAction() {
+        if (!this.id) {
+            return this.fail('params error');
+        }
         let pk = await this.modelInstance.getPk();
         let data = this.post();
         delete data[pk];
         if (think.isEmpty(data)) {
             return this.fail('data is empty');
         }
-        let rows = await this.modelInstance.where({ [pk]: this.id }).update(data);
-        return this.success({ affectedRows: rows });
+        let rows = await this.modelInstance.where({[pk]: this.id}).update(data);
+        return this.success({affectedRows: rows});
     }
 }
