+++ src/api/controller/users_thumbup.js
@@ -13,20 +13,20 @@ export default class extends think.controller.rest {
         super.init(http);
     }
 
+    /**
+     * before magic method
+     * @return {Promise} []
+     */
+    __before() {
+
+    }
+
 
     /**
      * 查询数据
      */
     async getAction() {
-        let data;
-        if (this.id) {
-            let pk = await this.modelInstance.getPk();
-            data = await this.modelInstance.where({ [pk]: this.id }).find();
-            return this.success(data);
-        }
-        let values = this.get();
-        data = await this.modelInstance.page(values.page, values.pageSize || 10).countSelect();
-        return this.success(data);
+        return this.fail();
     }
 
     /**
@@ -48,22 +48,18 @@ export default class extends think.controller.rest {
      * 删除数据
      */
     async deleteAction() {
+        if (!this.id) {
+            return this.fail('params error');
+        }
         let pk = await this.modelInstance.getPk();
-        let rows = await this.modelInstance.where({ [pk]: this.id }).delete();
-        return this.success({ affectedRows: rows });
+        let rows = await this.modelInstance.where({[pk]: this.id}).delete();
+        return this.success({affectedRows: rows});
     }
 
     /**
      * 更新数据
      */
     async putAction() {
-        let pk = await this.modelInstance.getPk();
-        let data = this.post();
-        delete data[pk];
-        if (think.isEmpty(data)) {
-            return this.fail('data is empty');
-        }
-        let rows = await this.modelInstance.where({ [pk]: this.id }).update(data);
-        return this.success({ affectedRows: rows });
+        return this.fail();
     }
 }
